refactoring pong code (in progress)
