+++ front-end/app/pong/normal/PongProvider.tsx
@@ -1,20 +1,9 @@
 "use client"
 
 import { WSClientType } from '@/helpers/wsclient';
-import { ReactNode, createContext, useState } from 'react';
+import { MutableRefObject, ReactNode, createContext, useRef, useState } from 'react';
 import { Dispatch, SetStateAction } from 'react';
-
-// TODO: Remove stuff I dont use
-// TODO: re-think this Provider, maybe is not needed
-
-// const initialBallState = {
-// 	position: { x: 0, y: 0 },
-// 	velocity: { x: 0, y: 0 },
-// };
-
-// const initialPaddleState = {
-// 	position: { x: 0, y: 0 },
-// };
+import { Mesh } from 'three'
 
 const initialPlayerState = {
 	name: "None",
@@ -35,13 +24,6 @@ const initialGameState = {
 };
 
 interface PongContextProps {
-	// ballState: {
-	// 	position: { x: number, y: number },
-	// 	velocity: { x: number, y: number },
-	// };
-	// paddleState: {
-	// 	position: { x: number, y: number },
-	// }
 	playerState: {
 		name: string,
 		color: number,
@@ -57,24 +39,23 @@ interface PongContextProps {
 		wsclient: WSClientType | null,
 		pause: boolean,
 	}
-	// updateBallState: Dispatch<SetStateAction<typeof initialBallState>>;
-	// updatePaddleState: Dispatch<SetStateAction<typeof initialPaddleState>>;
+	rightPaddleRef: MutableRefObject<Mesh>,
+	leftPaddleRef: MutableRefObject<Mesh>,
+	ballRef: MutableRefObject<Mesh>,
 	updatePlayerState :  Dispatch<SetStateAction<typeof initialPlayerState>>;
 	updateOpponentState :  Dispatch<SetStateAction<typeof initialOpponentState>> 
-	updateGameState :  Dispatch<SetStateAction<typeof initialGameState>> 
+	updateGameState :  Dispatch<SetStateAction<typeof initialGameState>>
 }
 
 const PongContext = createContext<PongContextProps>({} as PongContextProps);
 
 const PongProvider: React.FC<{ children: ReactNode, initialWsClient?: WSClientType | null }> = ({ children, initialWsClient }) => {
-	// const [ballState, setBallState] = useState( initialBallState );
-	// const [paddleState, setPaddleState] = useState( initialPaddleState );
-	const [playerState, setPlayerState] = useState( initialPlayerState ); // could be moved to another provider, like for all games
-	const [opponentState, setOpponentState] = useState( initialOpponentState );  // this too
-	const [gameState, setGameState] = useState({
-		...initialGameState,
-		wsclient: initialWsClient !== undefined ? initialWsClient : null,
-	});
+	const [playerState, setPlayerState] = useState( initialPlayerState );
+	const [opponentState, setOpponentState] = useState( initialOpponentState );
+	const [gameState, setGameState] = useState({ ...initialGameState, wsclient: initialWsClient !== undefined ? initialWsClient : null });
+	const rightPaddleRef = useRef<Mesh>(null) as MutableRefObject<Mesh>;
+	const leftPaddleRef = useRef<Mesh>(null) as MutableRefObject<Mesh>;
+	const ballRef = useRef<Mesh>(null) as MutableRefObject<Mesh>;
 
 	const updateGameState : Dispatch<SetStateAction<typeof initialGameState>> = ( newState ) => {
 		setGameState(prevState => ({
@@ -82,20 +63,6 @@ const PongProvider: React.FC<{ children: ReactNode, initialWsClient?: WSClientTy
 			...newState,
 		}));
 	}
-
-	// const updateBallState : Dispatch<SetStateAction<typeof initialBallState>> = ( newState ) => {
-	// 	setBallState(prevState => ({
-	// 		...prevState,
-	// 		...newState,
-	// 	}));
-	// };
-
-	// const updatePaddleState : Dispatch<SetStateAction<typeof initialPaddleState>> = ( newState ) => {
-	// 	setPaddleState((prevState) => ({
-	// 		...prevState,
-	// 		...newState,
-	// 	}));
-	// };
 
 	const updatePlayerState :  Dispatch<SetStateAction<typeof initialPlayerState>>  = ( newState ) => {
 		setPlayerState((prevState) => ({
@@ -114,8 +81,7 @@ const PongProvider: React.FC<{ children: ReactNode, initialWsClient?: WSClientTy
 	return (
 		<PongContext.Provider
 			value={{
-				// ballState, updateBallState, 
-				// paddleState, updatePaddleState, 
+				rightPaddleRef, leftPaddleRef, ballRef,
 				playerState, updatePlayerState,
 				opponentState, updateOpponentState,
 				gameState, updateGameState,
