+++ front-end/app/pong/normal/components/Border.tsx
@@ -1,11 +1,9 @@
-import * as THREE from "three"
-
 /**
  * Creates a Three.js mesh representing a 3D border within a 3D space.
  * @param position - The position of the boarder in the 3D space.
  * @returns A Three.js mesh representing the border.
  */
-const Border = ({ position }: {position: [number, number, number]}) => {
+export const Border = ({ position }: {position: [number, number, number]}) => {
 	return (
 		<mesh position={position}>
 			<boxGeometry args={[306, 4, 4]}/>
@@ -13,5 +11,3 @@ const Border = ({ position }: {position: [number, number, number]}) => {
 		</mesh>
 	);
 }
-
-export default Border
